Change from Spring WebFlux to MVC architecture

+++ src/main/kotlin/com/fisight/fisight/capital/Capital.kt
@@ -1,4 +1,4 @@
 package com.fisight.fisight.capital
 
 //TODO: add currency, wrap in money class
-data class Capital(val amount: Long)
+data class Capital(val amount: Double)
